use the 5D-downsampling code ...

+++ src/main/java/net/preibisch/mvrecon/process/n5api/N5ApiTools.java
@@ -450,8 +450,8 @@ public class N5ApiTools
 			final MultiResolutionLevelInfo mrInfo,
 			final MultiResolutionLevelInfo mrInfoPreviousScale,
 			final long[][] gridBlock,
-			final int currentChannelIndex,
-			final int currentTPIndex )
+			final long currentChannelIndex,
+			final long currentTPIndex )
 	{
 		final String dataset = mrInfo.dataset;
 		final String datasetPreviousScale = mrInfoPreviousScale.dataset;
